+++ LibTextFormat/core/Filters.lua
@@ -5,14 +5,20 @@ function LTF:ListFilters()
 end
 
 function LTF:RegisterFilter(name, fn)
-    local filters = self:ListFilters() or {}
-    filters[name] = fn
-    self:SaveToVars("filters", filters)
+  local filters = self:ListFilters() or {}
+  filters[name] = fn
+  self:SaveToVars("filters", filters)
 end
 
-function LTF:RegisterFiltersBulk(object)
-  for key, filter in pairs(object) do
+function LTF:RegisterFiltersBulk(object, version, selectiveOverrides)
+  version = version or self.defaultVersion
+  for key, filter in pairs(object[version]) do
     self:RegisterFilter(key, filter)
+  end
+  for ver, overridden in pairs(selectiveOverrides) do
+   if object[ver] ~= nil then
+    self:RegisterFilter(overridden, object[ver])
+   end
   end
 end