+++ LibTextFormat/filters/Items.lua
@@ -1,9 +1,9 @@
 LibTextFormat = LibTextFormat or {}
 local LTF = LibTextFormat
 
-LTF.Core = LTF.Core or {}
+LTF.Core = LTF.Core or { ["v1"] = {} }
 
-LTF.Core["item"] = function(itemId)
+LTF.Core["v1"]["item"] = function(itemId)
     if not itemId then return "" end
     local realStyle = LINK_STYLE_DEFAULT
     if style == "bracket" then