+++ LibTextFormat/filters/Utility.lua
@@ -1,17 +1,11 @@
 LibTextFormat = LibTextFormat or {}
 local LTF = LibTextFormat
 
-LTF.Core["plural"] = function(count, singular, plural)
-    if count == 1 then
-      return singular
+LTF.Core = LTF.Core or { ["v1"] = {} }
+
+LTF.Core["v1"]["plural"] = function(ctx, value)
+    if ctx.count == 1 then
+      return ctx.singular
     end
-    return plural or (singular .. "s")
-end
-
-LTF.Core["number"] = function(value)
-    return tonumber(value) or 0
-end
-
-LTF.Core["string"] = function(value)
-    return tostring(value or "")
+    return ctx.plural or (ctx.singular .. "s")
 end