+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2025.02.06
+- Protocol refactor
+- Added versioning
+- Removed encode and decode approach from protocols
+- Removed Math and String functions - this should probably be processed on the backend
+
 ## 2025.02.05
 - Fixes, added documentation